requirements.txt
streamlit>=1.30.0
openpyxl>=3.1.0
python-docx>=0.8.11
pandas>=2.0.0
playwright>=1.40.0
PyPDF2>=3.0.0

# V4: LLM providers (install what you need)
openai>=1.12.0          # OpenAI, Azure OpenAI, and Ollama (all use openai SDK)
boto3>=1.34.0           # AWS Bedrock (optional — skip if not using Bedrock)
